+++ src/Propulsion.Kafka/Producers.fs
@@ -7,13 +7,19 @@ open Serilog
 open System
 
 /// Methods are intended to be used safely from multiple threads concurrently
-type Producers(log : ILogger, clientId, broker, topic, ?customize, ?producerParallelism) =
+type Producer
+    (   log : ILogger, clientId, broker, topic, ?customize,
+        // Deprecated; there's a good chance this will be removed
+        ?degreeOfParallelism) =
     let cfg =
         KafkaProducerConfig.Create(
             clientId, broker, Acks.Leader,
             compression = CompressionType.Lz4, linger = TimeSpan.Zero, maxInFlight = 1_000_000,
             ?customize = customize)
-    let producers = Array.init (defaultArg producerParallelism 1) (fun _i -> KafkaProducer.Create(log, cfg, topic))
+    // NB having multiple producers has yet to be proved necessary at this point
+    // - the theory is that because each producer gets a dedicated rdkafka context, compression thread and set of sockets, better throughput can be attained
+    // - we should consider removing the degreeOfParallism argument and this associated logic unless we actually get to the point of leaning on this
+    let producers = Array.init (defaultArg degreeOfParallelism 1) (fun _i -> KafkaProducer.Create(log, cfg, topic))
     let produceStats = Streams.Internal.ConcurrentLatencyStats(sprintf "producers(%d)" producers.Length)
     let mutable robin = 0